[TRAFODION-2622] Left join with other_join_predicate on table alone is slow
This change is due to Hans Zeller.
The other join predicate in a hash join is evaluated before the hash join,
so that failing rows can be nulll instantiated directly. This avoids
a temporary join explosion when the join predicate matches most rows but
the other_join_predicate fails most rows. This change will be extended
to merge joins later.