[TRAFODION-2628] SQL engine work for Elasticity
It turns out that the compiler already kept a list of node ids,
not assuming that node ids are consecutively numbered 0...n-1.
However, this list was contained in lots of other data structures
that are dead code, coming from the time before Trafodion was
open-sourced. The change does two things: a) getting rid of obsolete
info and b) maintining a list of node ids in the NAClusterInfo
class.

This change does not include executor modifications. The executor
might work just fine, or we may need to add a fix (or optimization)
for cases where we pick a node id for a new process at runtime.