[TRAFODION-2628] Updates for elasticity
Replaced the list of CPU (node) numbers with an array, since it is
really a mapping array from logical to physical numbers.

Also changed some code in the adaptive segmentation code, but it is
likely that Adaptive Segmentation will not work (with or without my
change), this is not turned on by default and is not tested.