[TRAFODION-2628] Runtime work for elasticity
Adding a CPU array, similar to what's used in the compiler, to the CLI
globals. Make sure that when the compiler didn't specify a CPU (node)
id for an ESP, don't pick a node id that no longer exists.

Made a similar change for the SSMP.

Also got rid of some dead or now obsolete code.