[TRAFODION-2695] SSMP process ($ZSMxxx) sees too many opens from the master process
To avoid hanging deadlocks in the IPC layer a SQL Error warning 2026
is now returned when the application attempts to have more than 14
statements opened at the same time from a JDBC/ODBC connection

Maximum Nowait depth is 16. We need to spare 2 slots - One to send query
finished message and the other to get the response for the pending
nowaited query started message.