[TRAFODION-2738] Rowset buffer size during insert/upsert should be limited
Rowset insert includes UnpackRows operator in its query tree. UnPackRows
operator was allocating a larger memory based on rowset size and row size
to flow the data to its parent operator.

Introduced a CQD EXE_MEMORY_FOR_UNPACK_ROWS_IN_MB to limit the amount
of memory allocated in this operator. The default value is 100MB.

In addition, streamlined the NAMemory infrastruture to return NULL or
dump core when more than 2 GB of memory is allocated in one chunk.