[TRAFODION-2719] Check for truncation of character columns
When inserting into a UTF-8 character column with a fixed number
of characters, do a check whether the source string contains no
more than the allowed number of characters (just checking the
byte length is not enough). This bug does not show when using
sqlci, since sqlci does its own check before calling the
executor. It only happens when using JDBC or trafci.