[TRAFODION-2783] jdbc_test_cdh fails at times with type 2 JDBC driver dumping core
In case of Type 2 JDBC driver, the Trafodion SQL engine is a library that is dynamic
loaded into the process. Initialization of C++ static objects in the dynamic loaded libraries
are supposed to be done before dlopen returns. But the behavior seems to be nondeterministic when
there are multiple threads or when there are dependent static objects (An static object expects
another to be initialized before it). I think, the order of the initialization is not guaranteed
by the standard.

Refactored the code to initialize static object CharInfo::builtinCollationDB_ as part of the first
CLI call in a thread safe manner.