RMS infrastructure now reads the configured pid Max from /proc/sys/kernel/pid_max but limits it to a maximum of 128K. It can be overridden by "PID_MAX" environment variable in ms.env.
If the PID_MAX variable is set to 0, 64K is assumed to be configured pid max.
If the PID_MAX variable is set to less than 32K, then 32K is assumed to be configured pid max.

When a SQL process has a pin greater than the configured pid max, an error message
"Pid , is higher than the configured pid max <configured_pid>" is written to
the SQL log file and disables RMS for those processes. Node won't be brought down.

"Get Statistics for rms " shows the the configured pid max.

The environment variable to set the size of RMS shared segment size is now
changed to RMS_SHARED_SEG_SIZE_MB.

genms change for controlling pid_max and rms shared segment size