[TRAFODION-2974] Make event log reader and JDBC real TMUDFs
Roberta pointed out that we have two predefined UDFs, EVENT_LOG_READER
and JDBC, where the system administrator should have the ability to
control who can execute these functions.

To do this, these two UDFs cannot be "predefined" UDFs anymore, since
those don't have the metadata that's required for doing grant and
revoke.

Roberta also pointed out that the JDBC UDF should refuse to connect to
the T2 driver, for security reasons.

The fix leaves the predefined TMUDFs in place, for now, they will be
removed in R2.4 (see TRAFODION-2975).

The new "real" TMUDFs are in the "_LIBMGR_" schema, mostly for
convenience, as this schema has other UDFs that are created when
Trafodion is initialized.