[TRAFODION-2853] memory leak of ComDiagsArea in CmpContext heap of mxosrvr
The ComDiagsArea is allocated in many places and from different heaps in Trafodion, making it difficult to
detect the source of the leak. Hence, a different approach is taken to fix this issue.

Currently, ComDiagsArea is allocated in many places unconditionally even when SQL statement completes
execution without any error or warnings. Then it is deallocated. Changed this strategy and
allocate ComDiagsArea only when there is an error or warning while compiling or executing the SQL statement.

This should help the product in two ways
1) To reduce the pathlength.  The smaller query execution would benefit the most by chopping of few more microseconds.
2) Reduce the memory growth due to leaked ComDiagsArea