[TRAFODION-2205] invalid char at create schema when authorization name is long
There is code that converts the user ID to its username.  The buffer size
requested was not big enough to hold the return value.  In addition, the buffer
size check was returning an error but did not add an error to the ComDiags area;
therefore the returned error was ignored.
- Changed max len in calls to getAuthNameFromAuthID to the correct size.
- Set up the Diags area when buffer size is too small so the error is reported
  correctly, returns error 20235:

"Error returned while converting auth ID to auth name, status: xx ID: xx."

Also removed redundant methods:
-  getDBUserNameFromID - calls getAuthNameFromID instead
-  getDBUserIDFromName - calls GetAuthIDFromName instead