[TRAFODION-3009] Streamline error handling in Executor utility commands
Changes to avoid allocation of ComDiagsArea in some more places unless it is
needed to pass errors or warnings.

[TRAFODION-2917] Refactor Trafodion implementation of hdfs scan for text formatted hive tables

Disabling USE_LIBHDFS_SCAN by default. The new implementation of Hdfs Scan is now used to
scan the text type hive files.

Hdfs Scan is now stopped gracefully when the hive scan is cancelled. This avoids the random
core seen with new implementation.