[TRAFODION-2542] Grantor is not correct when granting privileges for a role
When granting privileges and the authorization ID is not the current user but
one of roles granted to the current user, then the "granted by" clause is
required.  In addition, the grantor of the privileges becomes the role specified
in the grant statement instead of the current user.

Added a CQD ALLOW_WGO_FOR_ROLES that will return an error if the user tries to
grant a privilege as a role.

Added error message (1194) when a component operation is not defined.

Added a check to not allow the WITH GRANT OPTION when granting privileges
to public