[TRAFODION-3225] Obscure cores seen in RMS and logger related code when Trafodion is stressed
Process de-registration is now logged to understand the spate of cores seen during
stress testing

Cleanup dangling semaphore incorrectly detected the problem due to pid recycling and
hence de-registered the process and its query fragment from the shared segment

When the RMS semaphore is held for more than 10 seconds, it is assumed to be in a
deadlock situation and hence the process holding the semaphore is core dumped for
further analysis.

Create a global logger instance instead of a function static variable
to help in debugging when problems with QRlogger is reported.