[TRAFODION-3260] SSMP may wait 3 seconds before handling requests
Encapsulated the changes in the commit 2ee272b325d021 to Ipc layer.

[TRAFODION-3274] At times sqlci or any other SQL process fails to come up and dumps core

It looks like there is a race condition in C++ main function prologue while
initializing the embedded global objects and the stdin, stdout and stderr file
descriptors.

File descriptor of value 2 is returned for epoll_create(). But, the error redirection
code possibly used this fd to redirect overriding the epoll fd.

This caused epoll_ctl to return EINVAL resulting in core dump of sql process.

Changed the global object gv_sock_ctlr to a global object pointer.