Optimizing and avoiding unnecessary function calls shown as hot spots in jprofile
java.net.Socket.setSoTimeout
org.trafodion.jdbc.t4.T4LoggingUtilities.log
java.net.Socket.getLocalSocketAddress

Now socket.read waits for the network timeout. It can be configured via a property networkTimeout or
call to connection.setNetworkTimeout. If the networkTimeout is not configured, a default value of 10 secs
is assumed.

The caller is expected to set the timeout depending upon the operation being performed like loginTimeout,
queryTimeout. Socket.read will wake up for every networkTimeout expiration and check if the timeout for the
operation hasn't exceeded. When it exceeds, SocketTimeoutException is thrown.

Changed the T4Statement and T4Restult classes to have "has a" relationship with T4Connection for
better understanding of the code. Earlier, it had "is a" relationship with T4Connection.