Optimizing and avoiding unnecessary function calls shown as hot spots in jprofile - part-II
java.lang.String.split and javal.lang.String.replaceAll

JDBC driver was parsing the sql string passed to it by the application to determine the type of
SQL statement and to check if the sql string has parameters. It doesn't parse the sql string anymore
and passes the sql string to SQL engine via mxosrvr. In case of PreparedStatement, the JDBC driver will
set the sqlStmtType based on the query type returned. In case of Statement, the JDBC driver will set the
sqlStmtType as SQL_TYPE_UNKNOWN.

Now, the PreparedStatement is always prepared for setting array values for parameter for all
statement types such as IUD and select. SQL engine will report an error if the array values are
not supported for any statement type.